:bug: Fixed music playing after removed from jukebox

+++ common/src/main/java/me/jaackson/etched/client/sound/DownloadProgressListener.java
@@ -9,7 +9,7 @@ public interface DownloadProgressListener {
 
     void progressStartRequest(Component component);
 
-    void progressStartDownload(int size);
+    void progressStartDownload(float size);
 
     void progressStagePercentage(int percentage);
